refactor(db): split channel membership store
